split public & registration fixture files
