+++ test/fixtures_registration.js
@@ -13,11 +13,6 @@
 // 09095555555: mother_father registration - father
 // cb245673-aa41-4302-ac47-1234567890 - id of mother with no msisdn
 
-// Identity roles - Public
-// 05059991111: heretofore unseen number
-// 05059992222: registered user - mother only registration, sms, pregnant, igbo
-// 05059993333: registered user -
-
 module.exports = function() {
     return [
         // 0: get identity 08080070007 by msisdn (to validate personnel_code)
@@ -3052,115 +3047,6 @@ module.exports = function() {
                     "operator": "cb245673-aa41-4302-ac47-00000000007",
                     "created_at": "2015-07-10T06:13:29.693272Z",
                     "updated_at":"2015-07-10T06:13:29.693298Z"
-                }
-            }
-        },
-
-        // ussd public
-        // 71: get identity 05059991111 by msisdn - no results
-        {
-            'request': {
-                'method': 'GET',
-                'params': {
-                    'details__addresses__msisdn': '+2345059991111'
-                },
-                'headers': {
-                    'Authorization': ['Token test_key'],
-                    'Content-Type': ['application/json']
-                },
-                'url': 'http://localhost:8001/api/v1/identities/search/',
-            },
-            'response': {
-                "code": 200,
-                "data": {
-                    "count": 0,
-                    "next": null,
-                    "previous": null,
-                    "results": []
-                }
-            }
-        },
-
-        // ussd public
-        // 71: create identity 05059991111
-        {
-            'request': {
-                'method': 'POST',
-                'headers': {
-                    'Authorization': ['Token test_key'],
-                    'Content-Type': ['application/json']
-                },
-                'url': "http://localhost:8001/api/v1/identities/",
-                'data': {
-                    "details": {
-                        "default_addr_type": "msisdn",
-                        "addresses": {
-                            "msisdn": {
-                                "+2345059991111": {}
-                            }
-                        }
-                    }
-                }
-            },
-            'response': {
-                "code": 201,
-                "data": {
-                    "url": "http://localhost:8001/api/v1/identities/cb245673-aa41-4302-ac47-9091111111/",
-                    "id": "cb245673-aa41-4302-ac47-9091111111",
-                    "version": 1,
-                    "details": {
-                        "default_addr_type": "msisdn",
-                        "addresses": {
-                            "msisdn": {
-                                "+2345059991111": {}
-                            }
-                        }
-                    },
-                    "operator": null,
-                    "created_at": "2015-07-10T06:13:29.693272Z",
-                    "updated_at": "2015-07-10T06:13:29.693298Z"
-                }
-            }
-        },
-
-        // ussd public
-        // 72: get identity 05059992222 by msisdn
-        {
-            'request': {
-                'method': 'GET',
-                'params': {
-                    'details__addresses__msisdn': '+2345059992222'
-                },
-                'headers': {
-                    'Authorization': ['Token test_key'],
-                    'Content-Type': ['application/json']
-                },
-                'url': 'http://localhost:8001/api/v1/identities/search/',
-            },
-            'response': {
-                "code": 200,
-                "data": {
-                    "count": 1,
-                    "next": null,
-                    "previous": null,
-                    "results": [{
-                        "url": "http://localhost:8001/api/v1/identities/3f7c8851-5204-43f7-af7f-005059992222/",
-                        "id": "3f7c8851-5204-43f7-af7f-005059992222",
-                        "version": 1,
-                        "details": {
-                            "default_addr_type": "msisdn",
-                            "addresses": {
-                                "msisdn": {
-                                    "+2348080020002": {}
-                                }
-                            },
-                            "receiver_role": "mother",
-                            "preferred_msg_type": "sms",
-                            "preferred_language": "igbo"
-                        },
-                        "created_at": "2015-07-10T06:13:29.693272Z",
-                        "updated_at": "2015-07-10T06:13:29.693298Z"
-                    }]
                 }
             }
         },
